Fix saving state of OpenPgpKeyPreference

+++ plugins/openpgp-api-library/src/main/java/org/openintents/openpgp/util/OpenPgpKeyPreference.java
@@ -242,7 +242,7 @@ public class OpenPgpKeyPreference extends Preference {
         public SavedState(Parcel source) {
             super(source);
 
-            keyId = source.readInt();
+            keyId = source.readLong();
             openPgpProvider = source.readString();
             defaultUserId = source.readString();
         }
